+++ rust/c509-certificate/examples/cli/main.rs
@@ -10,12 +10,12 @@ use asn1_rs::{oid, Oid};
 use c509_certificate::{
     attributes::attribute::Attribute,
     big_uint::UnwrappedBigUint,
+    cert_tbs::TbsCert,
     extensions::Extensions,
     issuer_sig_algo::IssuerSignatureAlgorithm,
     name::{Name, NameValue},
     signing::{PrivateKey, PublicKey},
     subject_pub_key_algo::SubjectPubKeyAlgorithm,
-    tbs_cert::TbsCert,
     time::Time,
 };
 use chrono::{DateTime, Utc};
